[feature] Support ... parameter for plotting and improve model name handling
Add ... parameter to most plotting functions passed to the underlying
base plotting functions.
Add a title to createModel function (roxygen warning).
Change the determination of the model name to no default and use the one
provided if present.
Make it so that aggregatedDirectPaths also accept a single set of paths.
Update the documentation.