Implement import/export/check of likelihood profiles.
Implement importProfiles, exportProfiles and checkProfiles to perform
the corresponding actions on the profile likelihood object (which has
more information than what is added to the MRAmodel by addPLinfos).
Add test for these functions.
Track the .rds for the profile likelihood tests.
Add documentations for the functions (in the same file)
Modify profileLikelihood a bit to return the profile list indexed by the
simplified path name and update the .rds accordingly.